Replace NavBar with base and specialisations
- NavBarBase is the new base NavBar
- UnAuthNavBar is a specialisation for unauthenticated users, e.g. on landing page, with no menu button
- DashboardNavBar is a specialisation for dashboard pages with no login and signup buttons
